requirements.txt
streamlit
torch
torchvision
pillow
numpy
gdown